+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Race range selector for fantasy league standings
 - Race range selector for fantasy league stats
+- Teammate H2H point to fantasy league stats
 
 ## [2.16.9] - 2025-03-10
 ### Added